requirements.txt
pyautogui~=0.9.53
pynput~=1.7.3
pandas~=1.3.5
pillow~=9.2.0
numpy~=1.21.5
matplotlib~=3.5.3
opencv-python~=4.6.0.66
pywin32~=302
mss~=6.1.0
pydirectinput~=1.0.4
pyqt5~=5.15.7
yaml~=0.2.5
pyyaml~=6.0
torchvision~=0.13.1
tqdm~=4.64.1
seaborn~=0.12.0
requests~=2.28.1
thop~=0.1.1-2209072238
tensorflow~=2.9.1